refactor: split four modules under the 1000-line cap, enforce with a test
The 1000-line cap CLAUDE.md documents was never enforced: poly's custom rules
are ast-grep pattern matches and cannot count lines, so the "rust-max-lines"
check never existed and four files had drifted past the cap. Split each along
real seams into sibling modules (comms/daemon.rs -> daemon_handlers.rs,
mcp/mod.rs -> state.rs, mcp/helpers_calls.rs -> helpers_calls_scan.rs,
mcp/types.rs -> types_web.rs), behaviour-preserving (moved code plus re-exports
and visibility widening; every file now under 1000 lines). tests/max_lines.rs
walks src/**/*.rs and fails cargo test on any over-cap file, so every CI leg
enforces it feature-agnostically.

Refs #19.
