+++ src/mcp/types.rs
@@ -837,120 +837,11 @@ pub(super) struct RecentCallView {
 }
 
 #[cfg(feature = "crawl")]
-#[derive(Debug, Deserialize, Serialize, schemars::JsonSchema)]
-pub struct WebScrapeParams {
-    /// Absolute http or https URL to fetch.
-    pub url: crate::url::Url,
-    /// When true (default), chunk + embed + write to LanceDB so the page is
-    /// reachable via `search_documents`. When false, fetch and return metadata
-    /// only — useful for previewing a URL before paying the embedding cost.
-    #[serde(default = "WebScrapeParams::default_index")]
-    pub index: bool,
-    /// LanceDB `scope` tag. Default `"web:<host>"`. Override to share a scope
-    /// across many hosts or to namespace per project.
-    #[serde(default)]
-    pub scope: Option<String>,
-}
-
+pub(super) use super::types_web::{
+    WebCrawlPageOutcome, WebCrawlResponse, WebMapEntry, WebMapResponse, WebScrapeResponse,
+};
 #[cfg(feature = "crawl")]
-impl WebScrapeParams {
-    fn default_index() -> bool {
-        true
-    }
-}
-
-#[cfg(feature = "crawl")]
-#[derive(Debug, Serialize)]
-pub(super) struct WebScrapeResponse {
-    pub url: String,
-    pub final_url: String,
-    pub status_code: u16,
-    pub content_type: String,
-    pub bytes: usize,
-    pub chunks_indexed: usize,
-    pub indexed: bool,
-    pub scope: String,
-}
-
-#[cfg(feature = "crawl")]
-#[derive(Debug, Deserialize, Serialize, schemars::JsonSchema)]
-pub struct WebCrawlParams {
-    /// Seed URL. The crawler follows links breadth-first from this page.
-    pub url: crate::url::Url,
-    /// Overrides the global `[crawl].max_pages` cap for this call only.
-    #[serde(default)]
-    pub max_pages: Option<u32>,
-    /// Overrides the global `[crawl].max_depth` cap for this call only.
-    #[serde(default)]
-    pub max_depth: Option<u32>,
-    /// LanceDB `scope` tag. Default `"web:<host>"` derived from the seed URL's
-    /// host. Every page indexed by this crawl uses the same scope so
-    /// `search_documents { scope: ... }` retrieves them together.
-    #[serde(default)]
-    pub scope: Option<String>,
-}
-
-#[cfg(feature = "crawl")]
-#[derive(Debug, Serialize)]
-pub(super) struct WebCrawlResponse {
-    pub seed_url: String,
-    pub pages_visited: usize,
-    pub pages_indexed: usize,
-    pub total_chunks: usize,
-    pub scope: String,
-    /// Per-page indexing outcomes — surfaced so an agent can tell which URLs
-    /// landed in LanceDB vs which were skipped (binary content, empty body).
-    pub pages: Vec<WebCrawlPageOutcome>,
-    /// Crawl-level error, if any (e.g. seed URL unreachable). Per-page errors
-    /// land in `pages[*].error` instead.
-    pub error: Option<String>,
-}
-
-#[cfg(feature = "crawl")]
-#[derive(Debug, Serialize)]
-pub(super) struct WebCrawlPageOutcome {
-    pub url: String,
-    pub status_code: u16,
-    pub chunks_indexed: usize,
-    pub indexed: bool,
-    pub error: Option<String>,
-}
-
-#[cfg(feature = "crawl")]
-#[derive(Debug, Deserialize, Serialize, schemars::JsonSchema)]
-pub struct WebMapParams {
-    /// Site to discover. Returns sitemap entries + linked URLs without
-    /// fetching their bodies.
-    pub url: crate::url::Url,
-    /// Cap the number of URLs returned. Default 100, max 1000. `total_urls` always
-    /// reports how many were discovered, so a capped response is never mistakable
-    /// for a complete one.
-    #[serde(default)]
-    pub limit: Option<u32>,
-}
-
-#[cfg(feature = "crawl")]
-#[derive(Debug, Serialize)]
-pub(super) struct WebMapResponse {
-    pub url: String,
-    /// Every URL discovered, whether or not it survived `limit`.
-    pub total_urls: usize,
-    /// `total_urls > urls.len()` — the caller is holding a page, not the site.
-    pub truncated: bool,
-    pub urls: Vec<WebMapEntry>,
-}
-
-#[cfg(feature = "crawl")]
-#[derive(Debug, Serialize)]
-pub(super) struct WebMapEntry {
-    pub url: String,
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub lastmod: Option<String>,
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub changefreq: Option<String>,
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub priority: Option<String>,
-}
+pub use super::types_web::{WebCrawlParams, WebMapParams, WebScrapeParams};
 
 pub use super::types_admin::{CacheClearParams, CacheGcParams, CacheStatsParams};
 pub(super) use super::types_admin::{CacheClearResponse, CacheGcResponse, CacheStatsResponse};
